+++ react/ProductListContext.tsx
@@ -1,27 +1,46 @@
 import React, { createContext, useContext, useReducer, FC } from 'react'
 
-interface State {
-  visibleProducts: any[]
+export interface Product {
+  productId: string
+  [key: string]: any
+}
+export interface State {
+  nextImpressions: Product[]
+  sentIds: Set<string>
 }
 
 type ReducerActions =
-  | { type: 'ADD_VISIBLE_PRODUCT'; args: { product: any } }
-  | { type: 'RESET_VISIBLE_PRODUCTS' }
+  | { type: 'SEND_IMPRESSION'; args: { product: Product } }
+  | { type: 'RESET_NEXT_IMPRESSIONS' }
 
-const ProductListStateContext = createContext({})
-const ProductListDispatchContext = createContext({})
+export type Dispatch = (action: ReducerActions) => void
+
+const DEFAULT_STATE: State = {
+  nextImpressions: [],
+  sentIds: new Set<string>(),
+}
+
+const ProductListStateContext = createContext<State>(DEFAULT_STATE)
+const ProductListDispatchContext = createContext<Dispatch>(action => {
+  console.error('error in dispatch ', action)
+})
 
 function productListReducer(state: State, action: ReducerActions): State {
   switch (action.type) {
-    case 'ADD_VISIBLE_PRODUCT': {
+    case 'SEND_IMPRESSION': {
       const { product } = action.args
+      let nextImpressions = state.nextImpressions
+      if (!state.sentIds.has(product.productId)) {
+        state.sentIds.add(product.productId)
+        nextImpressions = state.nextImpressions.concat(product)
+      }
       return {
         ...state,
-        visibleProducts: state.visibleProducts.concat(product),
+        nextImpressions,
       }
     }
-    case 'RESET_VISIBLE_PRODUCTS': {
-      return { ...state, visibleProducts: [] }
+    case 'RESET_NEXT_IMPRESSIONS': {
+      return { ...state, nextImpressions: [] }
     }
     default: {
       throw new Error(`Unhandled action type on product-list-context`)
@@ -30,7 +49,8 @@ function productListReducer(state: State, action: ReducerActions): State {
 }
 
 const initialState: State = {
-  visibleProducts: [] as any[],
+  nextImpressions: [],
+  sentIds: new Set(),
 }
 
 const ProductListProvider: FC = ({ children }) => {