+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `useProductImpression` hook.
+
+### Changed
+-  The context's state props' names and dispatch actions.
 
 ## [0.1.1] - 2020-01-03
 ### Changed